Add defaults for LinesStartAt1, ColumnsStartAt1 and PathFormat arguments of initialize request (#55)
The DAP spec includes that the default values for LinesStartAt1, ColumnsStartAt1 and PathFormat are nonzero values (true, true, and "path"). Set these values before decoding, so omitted values map to those and not to zero values.

Additionally, remove  `omitempty` flags from LinesStartAt1 and ColumnsStartAt1 to avoid translation of missing false values to default true values on the receiving end.