cmd/gentypes: sync at DAP @ 1.47
https://github.com/microsoft/vscode-debugadapter-node/tree/049665a465bf74e5f78d648469acbef95f4b16fd

According to the spec, RestartArguments uses a union type (oneOf) and Go does not have a union type.
Both LaunchRequestArguments and AttachRequestArguments are represented as map[string]interface{},
so we may attempt to map this to map[string]interface{}. But if the spec starts to include other
types in the union, this will be a problem. For now, use interface{}.

Added -u and -o flags, and use them with the go generate tool.