Include the raw message in the error
When an error occurs, there is currently no good way to access the bad
message, because the bytes have been consumed from the input stream and
the error does not include the raw message. This makes it hard to debug
and makes logs less useful, because just the message type/command is
often not enough.

This PR adds the raw bytes to DecodeProtocolMessageFieldError to make
this easier.

Signed-off-by: Cornelius Weig <22861411+corneliusweig@users.noreply.github.com>